organized scripts into a different files :green_book:
